add inheritance errors to the report
OCD-3152
