+++ chpl/chpl-resources/src/main/resources/errors.properties
@@ -345,6 +345,7 @@ resources.schemaFileGeneralError=There was an error getting the file requested.
 #ics
 ics.noInheritanceError=Lack of Inheritance Source
 ics.badIncrementError=Invalid Code increment; is %s, should be %s
+ics.gapListingError=New 2015 Gap Listing Post Cures Rule Effective Date
 
 #surveillance
 surveillance.reasonRequired=A reason is required to delete a surveillance.
