+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/scheduler/InheritanceErrorsReportDAO.java
@@ -1,34 +1,93 @@
 package gov.healthit.chpl.dao.scheduler;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import javax.persistence.Query;
+
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import gov.healthit.chpl.auth.user.User;
+import gov.healthit.chpl.dao.impl.BaseDAOImpl;
 import gov.healthit.chpl.dto.scheduler.InheritanceErrorsReportDTO;
+import gov.healthit.chpl.entity.scheduler.InheritanceErrorsReportEntity;
 import gov.healthit.chpl.exception.EntityCreationException;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 
 /**
- * Data access object for the inheritance_errors_report table.
+ * The implementation for InheritanceErrorsReportDAO.
+ * 
  * @author alarned
  *
  */
-public interface InheritanceErrorsReportDAO {
-    /**
-     * Retrieves all of the inheritance_errors_report records as a list InheritanceErrorsReportDTO objects.
-     * @return List<InheritanceErrorsReportDTO> all records
-     */
-    List<InheritanceErrorsReportDTO> findAll();
+@Repository("inheritanceErrorsReportDAO")
+public class InheritanceErrorsReportDAO extends BaseDAOImpl {
 
-    /**
-     * Marks all inheritance_errors_report records as deleted.
-     */
-    void deleteAll();
+    public List<InheritanceErrorsReportDTO> findAll() {
+        List<InheritanceErrorsReportEntity> result = this.findAllEntities();
+        List<InheritanceErrorsReportDTO> dtos = new ArrayList<InheritanceErrorsReportDTO>(result.size());
+        for (InheritanceErrorsReportEntity entity : result) {
+            dtos.add(new InheritanceErrorsReportDTO(entity));
+        }
+        return dtos;
+    }
 
-    /**
-     * Inserts the data in the InheritanceErrorsReportDTO into the inheritance_errors_report table.
-     * @param dto InheritanceErrorsReportDTO object populated with the data to be inserted.
-     * @throws EntityCreationException when the data could not be inserted
-     * @throws EntityRetrievalException when the newly created record cannot be retrieved
-     */
-    void create(List<InheritanceErrorsReportDTO> dto)
-            throws EntityCreationException, EntityRetrievalException;
+    @Transactional
+    public void deleteAll() {
+        List<InheritanceErrorsReportEntity> entities = this.findAllEntities();
+
+        for (InheritanceErrorsReportEntity entity : entities) {
+            if (!entity.getDeleted()) {
+                entity.setDeleted(true);
+                entityManager.merge(entity);
+                entityManager.flush();
+            }
+        }
+    }
+
+    @Transactional
+    public void create(final List<InheritanceErrorsReportDTO> dtos)
+            throws EntityCreationException, EntityRetrievalException {
+        for (InheritanceErrorsReportDTO dto : dtos) {
+            InheritanceErrorsReportEntity entity = new InheritanceErrorsReportEntity();
+            entity.setChplProductNumber(dto.getChplProductNumber());
+            entity.setDeveloper(dto.getDeveloper());
+            entity.setProduct(dto.getProduct());
+            entity.setVersion(dto.getVersion());
+            entity.setAcb(dto.getAcb());
+            entity.setUrl(dto.getUrl());
+            entity.setReason(dto.getReason());
+            entity.setDeleted(false);
+            entity.setLastModifiedUser(getUserId(User.SYSTEM_USER_ID));
+
+            entityManager.persist(entity);
+        }
+        entityManager.flush();
+    }
+
+    @Transactional
+    public void create(final InheritanceErrorsReportDTO dto)
+            throws EntityCreationException, EntityRetrievalException {
+        InheritanceErrorsReportEntity entity = new InheritanceErrorsReportEntity();
+        entity.setChplProductNumber(dto.getChplProductNumber());
+        entity.setDeveloper(dto.getDeveloper());
+        entity.setProduct(dto.getProduct());
+        entity.setVersion(dto.getVersion());
+        entity.setAcb(dto.getAcb());
+        entity.setUrl(dto.getUrl());
+        entity.setReason(dto.getReason());
+        entity.setDeleted(false);
+        entity.setLastModifiedUser(getUserId(User.SYSTEM_USER_ID));
+
+        entityManager.persist(entity);
+        entityManager.flush();
+    }
+
+    private List<InheritanceErrorsReportEntity> findAllEntities() {
+        Query query = entityManager.createQuery("from InheritanceErrorsReportEntity iere "
+                + "where (iere.deleted = false)",
+                InheritanceErrorsReportEntity.class);
+        return query.getResultList();
+    }
 }
