+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/InheritanceErrorsReportCreatorJob.java
@@ -1,14 +1,14 @@
 package gov.healthit.chpl.scheduler.job;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
-import java.util.Properties;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,9 +16,7 @@ import org.quartz.DisallowConcurrentExecution;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
+import org.springframework.core.env.Environment;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.context.support.SpringBeanAutowiringSupport;
@@ -34,6 +32,7 @@ import gov.healthit.chpl.exception.EntityCreationException;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.manager.CertifiedProductDetailsManager;
 import gov.healthit.chpl.util.ChplProductNumberUtil;
+import gov.healthit.chpl.util.ErrorMessageUtil;
 
 /**
  * Initiates and runs the the Quartz job that generates the data that is used to to create the Inheritance Errors Report
@@ -46,9 +45,7 @@ import gov.healthit.chpl.util.ChplProductNumberUtil;
 public class InheritanceErrorsReportCreatorJob extends QuartzJob {
     private static final Logger LOGGER = LogManager.getLogger("inheritanceErrorsReportCreatorJobLogger");
     private static final String EDITION_2015 = "2015";
-    private static final String DEFAULT_PROPERTIES_FILE = "environment.properties";
     private static final int MIN_NUMBER_TO_NOT_NEED_PREFIX = 10;
-    private Properties props;
 
     @Autowired
     private CertifiedProductSearchDAO certifiedProductSearchDAO;
@@ -63,7 +60,12 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
     private ListingGraphDAO listingGraphDAO;
 
     @Autowired
-    private MessageSource messageSource;
+    private ErrorMessageUtil errorMessageUtil;
+
+    @Autowired
+    private Environment env;
+
+    private Date curesRuleEffectiveDate;
 
     /**
      * Constructor to initialize InheritanceErrorsReportCreatorJob object.
@@ -73,7 +75,6 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
      */
     public InheritanceErrorsReportCreatorJob() throws Exception {
         super();
-        loadProperties();
     }
 
     @Override
@@ -83,16 +84,22 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
 
         LOGGER.info("********* Starting the Inheritance Error Report Creator job. *********");
 
+        curesRuleEffectiveDate = getCuresRuleEffectiveDate();
+        if (curesRuleEffectiveDate == null) {
+            return;
+        }
+
         inheritanceErrorsReportDAO.deleteAll();
         List<CertifiedProductFlatSearchResult> listings = certifiedProductSearchDAO.getAllCertifiedProducts();
         List<CertifiedProductFlatSearchResult> certifiedProducts = filterData(listings);
+        certifiedProducts = certifiedProducts.subList(0, 100);
+
         ExecutorService executorService = null;
         try {
             executorService = Executors.newFixedThreadPool(8);
 
             for (CertifiedProductFlatSearchResult result : certifiedProducts) {
-                CompletableFuture<Void> future = CompletableFuture
-                        .supplyAsync(() -> getCertifiedProductSearchDetails(result.getId()), executorService)
+                CompletableFuture.supplyAsync(() -> getCertifiedProductSearchDetails(result.getId()), executorService)
                         .thenApply(cp -> check(cp))
                         .thenAccept(error -> saveInheritanceErrorsReportSingle(error));
             }
@@ -127,7 +134,8 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
             item.setProduct(listing.getProduct().getName());
             item.setVersion(listing.getVersion().getVersion());
             item.setAcb(listing.getCertifyingBody().get(CertifiedProductSearchDetails.ACB_NAME_KEY).toString());
-            String productDetailsUrl = props.getProperty("chplUrlBegin").trim();
+            String productDetailsUrl = env.getProperty("chplUrlBegin").trim();
+            LOGGER.info("productDetailsUrl = " + productDetailsUrl);
             if (!productDetailsUrl.endsWith("/")) {
                 productDetailsUrl += "/";
             }
@@ -149,29 +157,11 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
         }
     }
 
-    private Properties loadProperties() throws IOException {
-        InputStream in = InheritanceErrorsReportCreatorJob.class.getClassLoader()
-                .getResourceAsStream(DEFAULT_PROPERTIES_FILE);
-        if (in == null) {
-            props = null;
-            throw new FileNotFoundException("Environment Properties File not found in class path.");
-        } else {
-            props = new Properties();
-            props.load(in);
-            in.close();
-        }
-        return props;
-    }
-
     private List<CertifiedProductFlatSearchResult> filterData(
             final List<CertifiedProductFlatSearchResult> certifiedProducts) {
-        List<CertifiedProductFlatSearchResult> results = new ArrayList<CertifiedProductFlatSearchResult>();
-        for (CertifiedProductFlatSearchResult result : certifiedProducts) {
-            if (result.getEdition().equalsIgnoreCase(EDITION_2015)) {
-                results.add(result);
-            }
-        }
-        return results;
+        return certifiedProducts.stream()
+                .filter(cp -> cp.getEdition().equalsIgnoreCase(EDITION_2015))
+                .collect(Collectors.toList());
     }
 
     private String breaksIcsRules(final CertifiedProductSearchDetails listing) {
@@ -189,8 +179,7 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
             // check if listing has ICS but no family ties
             if (hasIcs && (listing.getIcs() == null || listing.getIcs().getParents() == null
                     || listing.getIcs().getParents().size() == 0)) {
-                return messageSource.getMessage(new DefaultMessageSourceResolvable("ics.noInheritanceError"),
-                        LocaleContextHolder.getLocale());
+                return errorMessageUtil.getMessage("ics.noInheritanceError");
             }
 
             // check if this listing has correct ICS increment
@@ -208,26 +197,39 @@ public class InheritanceErrorsReportCreatorJob extends QuartzJob {
                 if (icsCode.intValue() != expectedIcsCode) {
                     String existing = (icsCode.toString().length() == 1 ? "0" : "") + icsCode.toString();
                     String expected = (expectedIcsCode < MIN_NUMBER_TO_NOT_NEED_PREFIX ? "0" : "") + expectedIcsCode;
-                    return String.format(
-                            messageSource.getMessage(new DefaultMessageSourceResolvable("ics.badIncrementError"),
-                                    LocaleContextHolder.getLocale()),
-                            existing, expected);
+                    return errorMessageUtil.getMessage("ics.badIncrementError", existing, expected);
                 }
             }
 
-            if (listing.getCertificationEdition().get("id").equals("3") && !hasIcs) {
-                if (listing.getCertificationResults().stream()
-                        .filter(cert -> cert.isGap())
-                        .count() > 0) {
-
-                    return "Found the new error!";
-                }
-
+            if (!hasIcs && doesGapExistForListing(listing) && isCertificationDateAfterRuleEffectiveDate(listing)) {
+                return errorMessageUtil.getMessage("ics.gapListingError");
             }
 
         } catch (Exception ex) {
             LOGGER.error("Could not compare ICS value " + icsCodePart + " to inherits boolean value", ex);
         }
         return null;
+    }
+
+    private boolean doesGapExistForListing(CertifiedProductSearchDetails listing) {
+        return listing.getCertificationResults().stream()
+                .filter(cert -> cert.isGap() != null ? cert.isGap() : false)
+                .count() > 0;
+    }
+
+    private boolean isCertificationDateAfterRuleEffectiveDate(CertifiedProductSearchDetails listing) {
+        return new Date(listing.getCertificationDate()).after(curesRuleEffectiveDate);
+    }
+
+    private Date getCuresRuleEffectiveDate() {
+        String dateFromPropertiesFile = env.getProperty("cures.ruleEffectiveDate");
+        LOGGER.info("cures.ruleEffectiveDate = " + dateFromPropertiesFile);
+        SimpleDateFormat sdf = new SimpleDateFormat("MM/dd/yyyy");
+        try {
+            return sdf.parse(dateFromPropertiesFile);
+        } catch (ParseException e) {
+            LOGGER.error("Could not parse: " + dateFromPropertiesFile, e);
+            return null;
+        }
     }
 }
